remove GenerationData
GenerationData seemed to obscure code function to me, so I've removed it. Raw data is passed around to jobs now, hopefully making it more obvious how the data is getting used.

+++ Assets/Scripts/Game/World/Blocks/BlockAction.cs
@@ -8,11 +8,11 @@ namespace Wyd.Game.World.Blocks
 {
     public struct BlockAction
     {
-        public int3 GlobalPosition { get; private set; }
+        public float3 GlobalPosition { get; private set; }
         public ushort Id { get; private set; }
 
-        public BlockAction(int3 globalPosition, ushort id) => (GlobalPosition, Id) = (globalPosition, id);
+        public BlockAction(float3 globalPosition, ushort id) => (GlobalPosition, Id) = (globalPosition, id);
 
-        public void SetData(int3 globalPosition, ushort id) => (GlobalPosition, Id) = (globalPosition, id);
+        public void SetData(float3 globalPosition, ushort id) => (GlobalPosition, Id) = (globalPosition, id);
     }
 }
